[Data oriented] factor out predeclare_struct_primitives into its own module
